Core dump on empty string array key that should exist
LBFactoryMulti.php sometime throws a notice for an array key that must
exist (ILoadBalancer::GROUP_GENERIC which is an empty string). The issue
only occurs on php7.2.

Attempt to create a coredump in case we reach that impossible case. Copy
pasted from Tim Starling at T229433#5436845.

Harness that behind a feature flag $wgT232613, set to boolean true to
enable. So we can enable it just on a few servers.

Bug: T232613
Change-Id: I7b2e5d6699d4ba2365c40c55e4f81b815c3f016e