Add more log and context for T232613 logging
In makeServerArray, the $groupLoads sometime lack the key
ILoadBalancer::GROUP_GENERIC when it should always be set by one the
caller.

In the coredump log, add some more context, notably the whole content of
the faulty $groupLoads, the value of the constant as well as wether it
is strictly identical to an empty string.

In the caller which actually set the key, add the same contextualized
log whenever the feature flag has been enabled (1/11 chance on
wikidatawiki, should not be too spammy?).

Bug: T232613
Change-Id: Ib8944f1a065b5d1983bb2050bf1409c0d8a88985