Ensure that the $target is a UserIdentity before calling ::getId()
AbstractBlock::parseTarget() may return a UserIdentity, a string, or null for
the target. It is possible for the method to return a string for a non-range
block if the IP is invalid. To prevent a fatal error, check the return type
rather than the block type.

Bug: T236425
Change-Id: I98dfcafe06b8621041cd4f72815b418e913eb1b0
(cherry picked from commit 6100226d3b3150b665ac14c2c214b5e8ffcf8b76)