Parsoid VRS: Add the Host header
While a MediaWiki instance, Parsoid/PHP is currently running as a
separate entity. All MW servers expect the Host header to be set in
order to be able to determine which wiki DB to select and work with.
Therefore, we are adding the Host headers to all requests issued to
Parsoid via VirtualRESTServiceClient.

Concretely, this change is needed for Flow and VE in private wikis to
work with Parsoid/PHP.

Bug: T229015
Bug: T229078
Bug: T229074
Change-Id: Ie14015350332a89bb4f9329ee0f4274eb2d745cb
(cherry picked from commit c45bf41a8d1d2cafa882dc715e6f72040704e0e1)